Fixed namespaces of previous refactoring
