+++ src/Accord.Bot/Helpers/Permissions/DiscordPermissionType.cs
@@ -1,4 +1,4 @@
-namespace Accord.Bot.Helpers
+namespace Accord.Bot.Helpers.Permissions
 {
     public enum DiscordPermissionType
     {
